requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
blinker==1.7.0
click==8.1.7
Django==4.2.11
djangorestframework==3.15.1
Flask==3.0.2
Flask-RESTful==0.3.10
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
ply==3.11
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
Werkzeug==3.0.1
zipp==3.18.1
